requirements.txt
# Database dependencies
psycopg2-binary
sqlalchemy>=1.4.0,<2.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# OpenAQ API client (for data fetching)
openaq>=0.5.0

# Environment variables
python-dotenv>=1.0.0

# Streamlit frontend dependencies
streamlit>=1.28.0
pydeck>=0.8.0
altair>=5.0.0